requirements.txt
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
idna==3.7
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.1
requests==2.31.0
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
